modified things based on review feedback
