+++ internal/plugin/config_plugin_test.go
@@ -10,7 +10,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/nginx/agent/v3/test/helpers"
+	"github.com/nginx/agent/v3/test/protos"
 
 	"github.com/nginx/agent/v3/internal/config"
 
@@ -69,9 +69,9 @@ func TestConfig_Process(t *testing.T) {
 		CorrelationID: correlationID,
 	}
 
-	configurationStatusProgress := helpers.CreateInProgressStatus()
+	configurationStatusProgress := protos.CreateInProgressStatus()
 
-	configurationStatus := helpers.CreateSuccessStatus()
+	configurationStatus := protos.CreateSuccessStatus()
 
 	tests := []struct {
 		name     string
@@ -112,7 +112,7 @@ func TestConfig_Process(t *testing.T) {
 				},
 				{
 					Topic: bus.InstanceConfigUpdateTopic,
-					Data:  helpers.CreateSuccessStatus(),
+					Data:  protos.CreateSuccessStatus(),
 				},
 			},
 		},
@@ -181,10 +181,10 @@ func TestConfig_Update(t *testing.T) {
 		CorrelationID: correlationID,
 	}
 
-	inProgressStatus := helpers.CreateInProgressStatus()
-	successStatus := helpers.CreateSuccessStatus()
-	failStatus := helpers.CreateFailStatus("error")
-	rollbackInProgressStatus := helpers.CreateRollbackInProgressStatus()
+	inProgressStatus := protos.CreateInProgressStatus()
+	successStatus := protos.CreateSuccessStatus()
+	failStatus := protos.CreateFailStatus("error")
+	rollbackInProgressStatus := protos.CreateRollbackInProgressStatus()
 
 	tests := []struct {
 		name               string
@@ -226,7 +226,7 @@ func TestConfig_Update(t *testing.T) {
 				},
 				{
 					Topic: bus.InstanceConfigUpdateTopic,
-					Data:  helpers.CreateRollbackSuccessStatus(),
+					Data:  protos.CreateRollbackSuccessStatus(),
 				},
 			},
 		},
@@ -249,7 +249,7 @@ func TestConfig_Update(t *testing.T) {
 				},
 				{
 					Topic: bus.InstanceConfigUpdateTopic,
-					Data:  helpers.CreateRollbackFailStatus("rollback failed"),
+					Data:  protos.CreateRollbackFailStatus("rollback failed"),
 				},
 			},
 		},
