+++ test/protos/instances.go
@@ -6,10 +6,98 @@
 package protos
 
 import (
+	"encoding/json"
 	"os"
+	"path"
+	"path/filepath"
+	"testing"
 
 	"github.com/nginx/agent/v3/api/grpc/instances"
+	"github.com/nginx/agent/v3/test/helpers"
+	"google.golang.org/protobuf/types/known/timestamppb"
+
+	"github.com/stretchr/testify/require"
 )
+
+const (
+	instanceID    = "aecea348-62c1-4e3d-b848-6d6cdeb1cb9c"
+	correlationID = "dfsbhj6-bc92-30c1-a9c9-85591422068e"
+)
+
+func CreateCacheFiles(t testing.TB, cachePath string, cacheData map[string]*instances.File) {
+	t.Helper()
+	cache, err := json.MarshalIndent(cacheData, "", "  ")
+	require.NoError(t, err)
+
+	err = os.MkdirAll(path.Dir(cachePath), os.ModePerm)
+	require.NoError(t, err)
+
+	for _, file := range cacheData {
+		helpers.CreateFileWithErrorCheck(t, filepath.Dir(file.GetPath()), filepath.Base(file.GetPath()))
+	}
+
+	err = os.WriteFile(cachePath, cache, os.ModePerm)
+	require.NoError(t, err)
+}
+
+func CreateInProgressStatus() *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_IN_PROGRESS,
+		Message:       "Instance configuration update in progress",
+		Timestamp:     timestamppb.Now(),
+	}
+}
+
+func CreateSuccessStatus() *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_SUCCESS,
+		Message:       "Config applied successfully",
+		Timestamp:     timestamppb.Now(),
+	}
+}
+
+func CreateFailStatus(err string) *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_FAILED,
+		Message:       err,
+	}
+}
+
+func CreateRollbackSuccessStatus() *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_ROLLBACK_SUCCESS,
+		Timestamp:     timestamppb.Now(),
+		Message:       "Rollback successful",
+	}
+}
+
+func CreateRollbackFailStatus(err string) *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_ROLLBACK_FAILED,
+		Timestamp:     timestamppb.Now(),
+		Message:       err,
+	}
+}
+
+func CreateRollbackInProgressStatus() *instances.ConfigurationStatus {
+	return &instances.ConfigurationStatus{
+		InstanceId:    instanceID,
+		CorrelationId: correlationID,
+		Status:        instances.Status_ROLLBACK_IN_PROGRESS,
+		Timestamp:     timestamppb.Now(),
+		Message:       "Rollback in progress",
+	}
+}
 
 func GetFileCache(files ...*os.File) (map[string]*instances.File, error) {
 	cache := make(map[string]*instances.File)
@@ -47,10 +135,10 @@ func GetFiles(files ...*os.File) (*instances.Files, error) {
 	return instanceFiles, nil
 }
 
-func GetFileDownloadResponse(path, instanceID string, content []byte) *instances.FileDownloadResponse {
+func GetFileDownloadResponse(filePath, instanceID string, content []byte) *instances.FileDownloadResponse {
 	return &instances.FileDownloadResponse{
 		Encoded:     true,
-		FilePath:    path,
+		FilePath:    filePath,
 		InstanceId:  instanceID,
 		FileContent: content,
 	}
